Reworked store/restore of original state.
1. Moved to TestCase to keep it DRY.
2. Removed store/restore for the Imagify's object property $secure_key. Why? When we restore the `$instance`, we are changing the object. The original secure_key remains on the original Imagify object we stashed in a property on the TestCase. Therefore, we don't need to store/restore the object's property. We only have to do it on static properties.
3. Added a TestCase to ImagifyUser tests to keep it DRY.
4. Removed test DocBlocks. Code is readable without it.
