+++ tests/Integration/inc/classes/ImagifyUser/isOverQuota.php
@@ -1,67 +1,31 @@
 <?php
+
 namespace Imagify\tests\Integration\inc\classes\ImagifyUser;
 
 use Brain\Monkey\Functions;
-use Imagify\tests\Integration\TestCase;
-
 use Imagify;
 use Imagify_User;
 
 /**
- * Tests for Imagify_User->is_over_quota().
- *
  * @covers Imagify_User::is_over_quota
  * @group  ImagifyAPI
  */
 class Test_IsOverQuota extends TestCase {
-	protected $api_credentials_config_file = 'imagify-api.php';
-
-	private $invalidApiKey = '1234567890abcdefghijklmnopqrstuvwxyz';
-
-	private $originalImagifyInstance;
-	private $originalUserInstance;
-	private $originalImagifyInstanceSecureKey;
-	private $originalApiKeyOption;
 	private $originalPreviousQuotaOption;
 
 	public function setUp() {
 		parent::setUp();
 
-		// Store previous state and nullify values.
-		$this->originalImagifyInstance = $this->resetPropertyValue( 'instance', Imagify::class );
-		$this->originalUserInstance    = $this->resetPropertyValue( 'user', Imagify::class );
-
-		if ( $this->originalImagifyInstance ) {
-			$this->originalImagifyInstanceSecureKey = $this->resetPropertyValue( 'secure_key', $this->originalImagifyInstance );
-		} else {
-			$this->originalImagifyInstanceSecureKey = null;
-		}
-
-		$this->originalApiKeyOption        = get_imagify_option( 'api_key' );
 		$this->originalPreviousQuotaOption = get_imagify_option( 'previous_quota_percent' );
 	}
 
 	public function tearDown() {
 		parent::tearDown();
 
-		// Reset state.
-		$modifiedImagifyInstance = $this->setPropertyValue( 'instance', Imagify::class, $this->originalImagifyInstance );
-
-		remove_filter( 'http_request_args', [ $modifiedImagifyInstance, 'force_api_key_header' ], IMAGIFY_INT_MAX + 25 );
-
-		$this->setPropertyValue( 'user', Imagify::class, $this->originalUserInstance );
-
-		if ( $this->originalImagifyInstance ) {
-			$this->setPropertyValue( 'secure_key', $this->originalImagifyInstance, $this->originalImagifyInstanceSecureKey );
-		}
-
-		update_imagify_option( 'api_key', $this->originalApiKeyOption );
+		// Restore the original option.
 		update_imagify_option( 'previous_quota_percent', $this->originalPreviousQuotaOption );
 	}
 
-	/**
-	 * Test Imagify_User->is_over_quota() should return false when couldn’t fetch user account data.
-	 */
 	public function testShouldReturnFalseWhenCouldNotFetchUserData() {
 		update_imagify_option( 'api_key', $this->invalidApiKey );
 
@@ -73,9 +37,6 @@ class Test_IsOverQuota extends TestCase {
 		$this->assertFalse( ( new Imagify_User() )->is_over_quota() );
 	}
 
-	/**
-	 * Test Imagify_User->is_over_quota() should return false when paid account.
-	 */
 	public function testShouldReturnFalseWhenPaidAccount() {
 		update_imagify_option( 'api_key', $this->getApiCredential( 'IMAGIFY_TESTS_API_KEY' ) );
 
@@ -94,9 +55,6 @@ class Test_IsOverQuota extends TestCase {
 		$this->assertFalse( $imagifyUser->is_over_quota() );
 	}
 
-	/**
-	 * Test Imagify_User->is_over_quota() should return false when free and not over quota.
-	 */
 	public function testShouldReturnFalseWhenFreeNotOverQuota() {
 		update_imagify_option( 'api_key', $this->getApiCredential( 'IMAGIFY_TESTS_API_KEY' ) );
 
@@ -113,9 +71,6 @@ class Test_IsOverQuota extends TestCase {
 		$this->assertFalse( $imagifyUser->is_over_quota() );
 	}
 
-	/**
-	 * Test Imagify_User->is_over_quota() should return true when free and over quota.
-	 */
 	public function testShouldReturnTrueWhenFreeOverQuota() {
 		update_imagify_option( 'api_key', $this->getApiCredential( 'IMAGIFY_TESTS_API_KEY' ) );
